fixed autocomplete collection discovery
